- Added Screen properties: isShareable, isDownloadable, url. - Study entity is restricted if !isShareable. - Annotations and screen result not downloadable if !isDownloadable. - Study URL shown as external link in StudyViewer (not yet added to ScreenViewer) - BoutrosAnnotationImporter now provides Boutros and Horn loginIDs and passwords. - WebDataAccessPolicy grants view access to non-shareable Screens is user is the lab head, study lead, or collaborator. - AnnotationValue.numericValue changed to double, to prevent loss of precision that would occur with fixed-precision BigDecimal (we only use the value for sorting in the database side, so a double is fine). - ReagentVendorIdentifier.reagentIdentifier changed to vendorIdentifier for consistency with Well.vendorIdentifier.
git-svn-id: https://forge.abcd.harvard.edu/svn/screensaver/trunk@1835 80c535d0-b110-0410-b4de-9279f94142c3

+++ src/edu/harvard/med/screensaver/model/screenresults/AnnotationType.java
@@ -197,7 +197,7 @@ public class AnnotationType extends AbstractEntity implements MetaDataType, Comp
     AnnotationValue annotationValue = new AnnotationValue(this,
                                                           reagentVendorIdentifier,
                                                           textValue,
-                                                          isNumeric && textValue != null ?  new BigDecimal(textValue) : null);
+                                                          isNumeric && textValue != null ?  new Double(textValue) : null);
     boolean result = _values.add(annotationValue);
     if (result) {
       return annotationValue;
